Extended and more consistent interface
Most importantly this commit adds generic functions that take Directory
as an argument. This simplifies use-cases where all directories have to
be determined, e.g. as part of a configuration or something. Type signatures:

    getDir :: Parameters -> Directory -> IO FilePath
    getDir' :: Directory -> Parameters -> IO FilePath
    getDir_ :: Directory -> IO FilePath

    getFileName :: Parameters -> Directory -> FilePath -> IO FilePath
    getFileName' :: Directory -> Parameters -> FilePath -> IO FilePath
    getFileName_ :: Directory -> FilePath -> IO FilePath

Another cool feature is that we now have three basic flavours of
functions. It's already indicated by the type signatures listed above.
There are functions that take Parameters, then Directory and then goes
the rest. Then there are functions that take it the other way around and
lastly there is a function that uses "def" for Parameters. There is no
need to list type signatures of those added functions, because it's
fairly obvious what they look like. Just go see the haddock.